Fix issues found with range selection
Fixed issues on the utilities dashboard with the date range selector
